+++ server/apps/orgsite/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class FooConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.site"
+    name = "apps.orgsite"
