+++ server/apps/orgsite/tests/test_views.py
@@ -3,7 +3,7 @@ from rest_framework import status
 from rest_framework.reverse import reverse
 from rest_framework.test import APIClient, APIRequestFactory, force_authenticate
 
-from apps.site.views import SiteDetailsView
+from apps.orgsite.views import SiteDetailsView
 
 
 class TestSiteListView:
