+++ server/manifest/services/manifest.py
@@ -4,7 +4,7 @@ from typing import Literal, Optional
 from django.db import transaction
 from django.db.models import Q, QuerySet
 
-from apps.site.models import Site
+from apps.orgsite.models import Site
 from manifest.models import Manifest
 from manifest.services import EManifest, EManifestError, TaskResponse
 from manifest.tasks import save_to_emanifest as save_to_emanifest_task
