+++ server/org/admin.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
 from guardian.admin import GuardedModelAdmin
 
-from apps.site.models import Site
+from apps.orgsite.models import Site
 from org.models import Org, OrgAccess, OrgGroupObjectPermission, OrgUserObjectPermission
 
 admin.site.register(OrgAccess)
